+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - **helm**: Fix `vendor --prune` deleting charts with a custom directory
   **[#717](https://github.com/grafana/tanka/pull/717)**
+- **helm**: Add validation at vendoring time for invalid chart names
+  **[#718](https://github.com/grafana/tanka/pull/718)**
 
 ## 0.22.0 (2022-06-03)